chore: split quantel files
